Include '.' in result returned from FileOperations::getExtension()
Before FileOperations::getExtension() on e.g. "file.ext" returned just
   "ext"
now it returns
   ".ext"
When passed a filename without extension both the new and the old version return
an empty string.

Currently there's only one caller of this function in openMSX and for that one
this difference doesn't really matter (maybe the old behavior is slightly
simpler). But this change is motivated by some (possibly) upcomming change
(related to sound channel recording).

In general the new behavior is easier to work with. e.g.
   FileOperations::stripExtension() + FileOperations::getExtension()
always results in the original file, whether it had an extension or not. With
the old behavior you conditionally had to insert a '.' character inbetween. And
handling filenames that end in a '.' character required even more checks.

The path class in the C++ filesystem TS
   http://en.cppreference.com/w/cpp/experimental/fs/path
also handles file extensions in this 'new' way.