Support partial (cancelled) WD2793 write sector/track commands
Apparently these are used in DSK-PRO to write sectors with CRC errors. See this
thread for more details:
   https://www.msx.org/forum/msx-talk/software/pdi-to-dmk-using-dsk-pro-104-with-openmsx

Supporting partial write commands was actually relatively easy. We already
emulated (most) writes with the correct timing. Though before this patch when a
write command was cancelled half-way, none of the writes were actually written
to disk. The reason for this is the following:

The WD2793 emulation always operates at the raw track level. Though usually
openMSX is used with .dsk disk images, those can only represent sector
information instead of the full track information. So at some point we need to
convert one format to the other. This conversion is not terribly expensive, but
also not cheap. So we don't do this after every written byte, but only once
after the full sector/track is written. Though before this patch when the write
command was cancelled we didn't do this conversion at all.