Oops, the last FDC changes broke .dsk images
The write sector command is now emulated more accurately: apart from the actual
sector data (and CRC) it will also rewrite part of the synchronization and
marker bytes right before and after the sector. While writing those it's possible
that existing 'special' bytes (bytes with another encoding as the normal 0-255
bytes) are overwritten. So we check for that and possibly erase those from the
RawTrack::idam table.

But what I didn't realize was that the DSKDiskImage class uses the same write()
routine. And because of the order in which DSKDiskImage calls addIdam() and
write() the idam markers were set but then immediately erased again.

I should have caught this mistake during testing. I did test .dsk but mostly
.dmk images during development (because this stuff is only relevant for .dmk).
But it seems I didn't retest .dsk after enhancing the write() routine :(