Move track-buffer from FDC to disk-drive.
Before this patch our FDC emulation (WD2793 or TC8566AF) reads a track from
disk into a buffer. It then manipulates that buffer. And (in case of a write
command) writes that buffer back to disk.

This is not how a real FDC operates. It does not have a buffer for the track.
Instead it reads/writes bytes while the disk is rotating. This patch moves the
code more towards that model. It's not yet exactly like the real hardware
because the FDC code still keeps track of the position in the track. More
specifically it requests to read/write a byte at a certain position in the
track instead of basing everything on the timing of the read/write request. So
read/write at a specific time when the desired byte is rotated below the disk
head. Our emulated timing is already pretty good, but not yet detailed enough
for a fully time-based track-addressing.

This patch only restructures the code: basically it changes the location of the
track-buffer in the code (from FDC to RealDrive) and then tweaks the interface
between FDC and Drive. It shouldn't have any functional changes.

The main motivation for this patch is as a preparation for the next patch.
There we will couple write-enable with motor-on. But currently the FDC isn't
aware of the motor status. And I don't want to make it aware of it because:
 - It's not trivial when one FDC controls multiple drives.
 - Also the real FDC (at least WD2793) isn't aware of the motor status.
With this refactoring we can let the drive make the write-enabled decisions
because that one does know the motor status. See next patch for more details.