Don't allow reading/writing from/to disk when disk isn't rotating
We already checked this, but only at the start of a command. Not when the motor
is turned off in the middle of a command.

Apparently dsk-pro does this to avoid writing the CRC bytes on a write-sector
command. See cbsfox' comment in this forum post:
   https://www.msx.org/forum/msx-talk/software/pdi-to-dmk-using-dsk-pro-104-with-openmsx?page=1

At first sight this is counter-intuitive to me. In the current emulation when
you turn the drive motor on/off the disk immediately starts/stops rotating at
300rpm. Though in the real world the disk has inertia and it takes some time
before the disk really stops rotating after the motor is turned off. Still
according to cbsfox turning the motor off reacts faster than aborting the FDC
command. I assume that in real hardware the drive's write-enable signal is
ANDed with the motor-on signal. So that writes are immediately disabled when
the disk starts slowing down.