Check on reset when doing record_chunks_on_framerate_changes
When a reset occurs, the framerate may change without a VDP write, as the VDP
itself is reset. Reset usually happens mid-frame, so that will definitely mean
a frame rate change with desync. So this is also by definition a frame rate
change and should result in a new video file.