Make DeltaBlock-compression handle changing blob-sizes
This fixes an assert when using certain .dmk disk images.

When creating an in-memory savestate, we delta-compress the binary blobs (e.g.
the content of the RAM). Because most often only a fraction of the blob has
changed between two consecutive savestates, and we gain both speed and reduce
memory usage by only storing the difference.

Before this patch, the code didn't expect the size of the blob to change from
one savestate to the next. (E.g. the size of the RAM always remains the same).
But it can happen when serializing a RawTrack class. Typically the size of a
track is 6250 bytes, but when using .dmk images that were extracted from a real
disk, there can be some variation in the exact size.

Fixed by identifying the blobs not only on their location in memory (a pointer)
but also on their size. That is if a blob changes in size we now see it as a
separate entity, so we don't attempt to calculate difference with the previous
smaller/bigger blob.