Add padding bytes to .rom files
Before this path the constructor of RomBlocks (the base class of most ROM
mappers) would check whether the .rom file was an integer multiple of the block
size (e.g. 8kB). If not it would refuse to use that ROM. Now we automatically
pad the ROM with 0xFF bytes.

The motivation for this change is that sometimes you find .rom files that are
not an exact multiple of 8kB or 16kB. Often those .rom files are the output of
an assembler/compiler rather than a dump from an actual ROM. Emulators like
blueMSX can directly handle such .rom files, but in openMSX you got an error
message.