'Fix' copy-protection detection in cracked disk version of Nemesis
The problem was that the emulated WD2793 read-track command was too 'clean'. It
returned the proper sequence 'A1 A1 A1', while a real WD2793 systematically
returns a 'wrong' value for the initial A1 byte. So it returns a sequence like
'14 A1 A1' or 'C2 A1 A1'. See comments in the code for a lot more details.

After this patch the game does start, but it soon crashes. So there might be
something else still wrong. Needs more investigation.