Fix emulation of Sunrise copy protection
My last patch, the applyWd2793ReadTrackQuirk() stuff, broke it.

Now we only apply this quirk in case of an actual read-track command. Usually
it's fine to apply this always, but I forgot about the overlapping sectors that
are used for the Sunrise disk copy protection.

Anyway, this is further evidence that this read track stuff is really a quirk
and that the data on disk really has 3 consecutive "A1" bytes.

This implementation is also cleaner because it keep the hack more localized to
the WD2793 code. E.g. the DMKDiskImage and SectorBasedDisk classes no longer
need to be aware of it.