Do a full rotation is next sector == current sector.
Found by running AurorA's Ruby & Jade and Exor games. These disks have a copy
protection with a track with only a single sector on it. The next sector is
then a full rotation away, but a bug in openMSX meant it would not rotate at
all and that would take 0 time. But finding the sector is tried for 5
revolutions... so that would take forever. In fact, openMSX hung up here.

The solution is to let openMSX do a full rotation if it should move to the same
location, instead of no rotation. Thanks to Wouter for finding out this root
cause.