Improve WD2793 read-address command
The WD2793 read-address command didn't report sector headers that did not have
an accompanying data block (either because the data block is missing or because
the header itself has a CRC error).

Discovered while dumping 'Shrines of Engima' with 'DMK creator' on openMSX.
There's a sector-header with CRC error on this disk. Because of this bug this
header was not included in the dumped. Unfortunately the game itself doesn't
check for the presence of this header, so the wrong dump appeared to work fine.