Fix head load timing in WD2793 emulation
TL;DR:
  Removed 1ms delay from type-2 and type-3 WD2793 commands.

Summary of the WD2793 datasheet related to head-load stuff:
- At the start of a read or write command, the WD2793 will activate the HLD pin
  (pin 28).
- The disk drive should reacts to this by 'engaging' the drive head. When this
  is finished the drive should respond by activating the HLT input pin (pin
  40). The datasheet lists a typical response time of 30-100ms for most disk
  drives.
- The WD2793 can only start reading/writing after this response.
- The HLD signal remains active after the read/write command has finished. So
  the next command doesn't suffer from this delay.
- HLD is automatically deactivated when the WD2793 has been idle for 15 index
  pulses (15 disk revolutions, at 300rpm that's 3 seconds).
- HLD can also be explicitly activated or deactivated by the seek commands (the
  h-bit in the command byte).
- The HLT response can be tested via bit 5 in the status register (it's
  actually the logical AND between HLD and HLT).

Though when I checked the schematic of several MSX machines (I checked Philips
NMS8245, Sony HB-G900P and the equivalent pins on the TC8566AF in a Panasonic
FS-A1GT) I see that the HLD output pin is unconnected and the HLT input pin is
directly connected to +5V. This means that the disk drive doesn't care about
HLD and it permanently replies "head-is-loaded". IOW there is no delay for head
loading.

Side note (mostly guessing here):
  The WD2793 datasheet seems to be written for 'old' 8" disk drives, not for
  'new' 3.5" drives. Possibly these 'new' drives don't need an explicit head
  load signal anymore???

Anyway, when I wrote the original WD2793 emulation in openMSX 15(!) years ago,
I had no idea about this head loading stuff. And there was this TODO in the
code:

      // TODO wait till head loaded, I arbitrarily took 1ms delay

But actually there's no delay at all. Moreover openMSX wrongly added this delay
on every read/write command, not only on the first command after an idle
period.

A 1ms delay corresponds to a disk rotation over ~30 bytes (there are ~6250
bytes in a track and the disk spins at 300rpm). Usually this 1ms delay is
masked by the much larger rotational seek delay, but not always.

One situation where this delay becomes apparent is in tools like Dmk Creator or
READ-DMK when those tools wait for the index pulse and then execute a (series
of) read-id command(s). If the first sector is located too close to the index
pulse it can't be read because of this emulation bug.

As a bonus this patch also adds more accurate emulation of the HEAD-LOADED bit
in the WD2793 status register.