Tweak/fix rotational delay for disk emulation
The method RealDrive::getNextSector() will rotate the disk till the next sector
header has passed below the disk head. Such a header is a sequence of 'A1 A1 A1
FE' bytes (A1 bytes with a 'special' encoding so that it cannot be confused
with a similar sequence that is part of normal sector data).

Before this patch that method would be happy to rotate the disk over just 1 or
2 bytes if the disk was already positioned very close to (the end of) such a
sequence. But in reality the FDC actually has to 'see' these 4 bytes before it
can conclude there was a sector header.

So this patch fixes this method to always rotate over at least 4 bytes.

This fixes a crash in Dmk Creator 5.1:
 - The program starts a read-address command.
 - Next it has some (partially intentional) delay of about 500 Z80 cycles.
 - Only then it reads the first data byte.
Before this patch, because of the too short rotational delay, openMSX would
report a LOST-DATA error in some cases (which eventually resulted in a hang).
After this patch that problem is fixed.