Implemented RomBlock<BANK_SIZE>::peekMem()
This fixes a bug reported by Laurent Halter in SofaROM. See here for some
more info:
   https://www.msx.org/forum/msx-talk/software/sofarom-bug-report-thread?page=6

After some debugging the problem turned out to be a missing implementation of
the peekMem() method in the RomBlock<BANK_SIZE> class.

Usually it's not a problem if this method is not implemented. The base class
MSXDevice has a default implementation with a safe fallback (= show no useful
information in the debugger in some cases). Though RomPanasonic::readMem() was
implemented in terms of Rom8kBBlocks::peekMem(), so for that class it's crucial
to have a complete peekMem().