Fixed issue#1080, crash with z80_freq unlocked and VDP-delay
The MSXMatsushita device implements a delay when accessing the VDP too fast.
There was a subtle bug in the implementation of this delay. It comes down to
the difference between Clock::advance() and Clock::reset():

- Both methods will set the clock's current time to some later time, but
- Clock::reset() simply sets the time to the given (future) time. At that point
  the clock starts ticking with a new clock edge.
- Clock::advance() (conceptually) keeps ticking the clock until the latest
  clock edge that is not past the given time.
- IOW reset() changes the 'phase' of the clock but advance() does not.
- Note that Clock::advance() possibly stops a fraction of a clock cycle before
  the requested time.

The intention of the MSXMatsushita code was probably to use reset() instead of
advance(). And that indeed fixes the crashes (at least I can't reproduce them
anymore).

This also explains why the crash only triggered with a much higher Z80 clock
freq than the default 3.5MHz: All Z80-IO actions are performed at clock-edges
(openMSX emulates an idealized time), when the Z80 runs at 3.5MHz those edges
exactly align with the edges of the MSXMatsushita clock. IOW when both clocks
run at the same speed (and are initially aligned, that's the case here) then
there's no difference between reset() and advance(). (I'm skipping over some
details, but you get the idea).