Fix issue #1069, tab-completion show internal procs
The problem was that e.g. "vgm[tab]" showed too many internal details. Next to
the desired list of commands:
  vgm_rec   vgm_rec_end   vgm_rec_next
it also showed a much longer list with stuff like:
  vgm::find_all_scc       vgm::update_time   vgm::write_opl4_address_1
  vgm::little_endian_32   vgm::vgm_rec       vgm::write_opl4_address_2
  ... 24 internal procs in total ...
Usually the user is not interested in this internal stuff.

This patch tries to improve this. Now "vgm[tab]" shows this list:
  vgm::   vgm_rec   vgm_rec_end   vgm_rec_next
So next to the 3 public vgm* command there is also "vgm::". So it shows there
is a "vgm::" namespace, but it doesn't show the internals anymore.

If you complete "vgm::[tab]" then you do get to see all the procs inside the
"vgm::" namespace. So in a way it now works like how filename/directory
completion works.