Fixed swapped row/column in TouchPad transformation matrix
After upgrading to gcc-7 the function Touchpad::parseTransformMatrix() seemed
to be miss-compiled. That function has 2 nested for-loops. But the generated
assembly code only contained one (infinite) loop. gcc-5 generated code with two
(finite) loops, just like you'd expect from the c++ code.

After some investigation it turned out that the c++ code actually contained
undefined behavior. We were accessing an array out of bounds. So the newer
compiler was not wrong, it just 'exploited' the undefined behavior.

More in detail: the code had a 2-by-3 matrix (2 rows, 3 columns) of which we
tried to access the 3rd row because I swapped the row and the column indices.
If you analyse which elements the gcc-5 generated code accesses, it turns out
they all fall within the memory of the 6 elements of the full matrix (so no
out-of-bound matrix access). It even turns out that the initial default value
for the matrix
   [ 256   0   0 ]
   [   0 256   0 ]
ended up correctly in memory with the exception of the 3,2 element which was
left uninitialized (but because it should be zero, there was a good chance even
that ended up correct). Maybe this explains why this bug went unnoticed.

To prevent similar bugs in the future I added bounds-checking in the matrix
code (though only enabled in a DEBUG build).