Make non-power-of-two mappers wrap in a more likely way
While I don't have the hardware to test this, it seems very unlikely
that for example a 640 kB mapper would wrap differently when accessed
in its second megabyte of address space compared to its first, which
is what happened before this commit. I expect address decoding
hardware to ignore all address lines that it does not need.