Improve MapperIO debuggable, to be more compatible with the old one
Before this commit, the MapperIO debuggable was implemented as an
I/O peek, which means that both the readback mask and the leading
ones from the mappers are applied to it. This is different from
the old debuggable, which just returned the last written value.

Now the MapperIO debuggable will take the maximum over the stored
segment number of all mappers. While this is not exactly the same
as the written value, it only differs if power-of-two size limits
cause the value to wrap, which doesn't happen often in practice:
the only scenario I know of is a memory search routine that wants
to detect the number of segments in a mapper.