Add emulation of the SN76489 sound chip
This chip is used in the Musical Memory Mapper, Franky and PlaySoniq
extensions for MSX, as well as in other Z80-based systems such as
early Sega consoles and the ColecoVision.

I don't have such hardware myself, so I had to rely on references,
which I listed in the header. Unfortunately, these are not complete
and sometimes contradict each other, so I had to apply some educated
guessing as to how the actual hardware is likely to work.