Add Musical Memory Mapper extension
This is a 1MB memory mapper with write protect feature and an SN76489
sound chip. It can be used to run MSX1 and Sega SG-1000 ROM images,
using MSX-DOS loader tools. It can use be used with Grauw's VGMPlay
or as a normal memory mapper.

I don't have the real hardware, so I implemented this using the info
from the MMM documentation, which is pretty detailed. I also used the
ROM loader tools to test the emulation:

https://www.msx.org/downloads/rom2mmm-ver-08
https://www.msx.org/forum/msx-talk/software/sg-1000-game-loader-msx-dos