Add ability to specify fixed slot map in XML
The ColecoVision and Sega SG-1000 doesn't have a slot structure
like the MSX and SVI have, but openMSX is designed around that,
so we have to create a fake slot structure.

This commit adds a <slotmap> tag to set the initial slot selection.
In systems without a way to change the slot selection, the initial
selection will remain active all the time.

Currently only primary slots can be selected in this way. It would
not be hard to allow secondary slots to be selected as well, but as
this is not needed for any machine that we have plans to support,
I'm not going to implement it at this time.