Remove duplicate code for mapping PPI port address to method
Every class using the I8255 included the same trio of switch
statements to map the Z80 port address to an I8255 method,
except for SVIPPI which had a small variation that I consider
to be a bug.

The solution is to have the I8255's read/peek/write methods
accept a port address (0-3) and put the switch statements inside
the I8255 class.