Move handling of keyboard configuration into Keyboard class
The code for handling the configuration was copy-pasted from MSXPPI
to SVIPPI. Now it resides in the Keyboard class instead.

If we ever want to do proper unit testing with dependency injection,
the DeviceConfig object is easy enough to construct in a unit test,
so I think removing all the fine-grained constructor parameters is
not going to be a problem.