Tweak scan_mismatch() in DeltaBlock.cc
scan_mismatch() is used during snapshot creation to compare the current with
the previous state and only store the difference in the snapshot.

Strictly according to the C++ standard the old implementation of
scan_mismatch() triggered undefined behavior because it could read a (partial)
word beyond the end of a memory buffer. Though we could justify that in all
practical implementations this was never a problem (see comments in old code
for details).

Nevertheless this (harmless) undefined behavior was reported by tools like
"address sanitizer". And it was reported often enough to swamp potential other
real problems. So this patch "fixes" it.

The new implementation never reads beyond the end of a buffer, but as a
trade-off it runs a little slower while scanning the last few bytes in the
buffer. "Slower" as in "on average it executes more x86 instructions". Though
the new code is shorter (less static instructions), so there's less pressure on
the instruction cache and potentially also on the branch-predictor. I couldn't
actually reliably measure a real-time speed difference between the two.

The new routine also takes advantage of SSE2 instructions (when available). For
large memory buffers (e.g. 2MB MoonSound sample RAM that rarely changes) this
does result in a measurable speedup (~25% faster).