Fixed crash on invalid slot specification
The slot names 'a'..'p' in stuff like
  <primary slot="b">
    <secondary slot="b">
are used internally to handle stuff like '-extb' to insert an extension in a
specific external slot. Both the <primary> and the <secondary> tag must specify
the same name.

Normally these names 'a'..'p' should not be used in manually written machine
and/or extension configuration files. Though of course fuzz-testing doesn't
care about this and therefor triggered a crash. Now added a check for it.