Fixed crash on invalid <window> tag inside <rom>
The <window> tag inside a <rom> specifies that only part of the full file
content should be used as the rom content. Like in this example:
    <rom>
      <sha1>...</sha1>
      <window base="0x28000" size="0x4000"/>
    </rom>
We check that the specified window falls inside the file. Though because of an
error in the order of initialization we trigger an assertion when this
situation occured. Fixed now.

Found while running afl-fuzz on openmsx machine configs.