Fixed assertion on invalid XML
The fuzzer sometimes produced configurtaion files that contain the following
fragment
   <!DOCTYPE  SYSTEM >
Normally the DOCTYPE looks like this
   <!DOCTYPE msxconfig SYSTEM 'msxconfig2.dtd'>

Before this patch our parser triggered an assertion because it tried to read a
character past the end of a string_ref when trying to figure out whether single
or double quotes were used after the SYSTEM keyword.

In a non-assert build this was harmless, but we should fix it anyway.