Fixed crash on invalid VDP/MSXMatsushita combination
The MSXMatsushita device wraps the VDP IO ports (to introduce some delay when
those ports are accessed). It expects to find those ports on address 0x98-0x9B.
And there was code to check for that. Though when this error triggers, some
state was not properly restored and that triggered a crash later on.

Fixed by replacing the (un)wrap_IO() methods in MSXCPUInterface with safer
replace_IO() methods.

Found while running afl-fuzz on openmsx machine configs.