Only create RomDatabase on first use
During fuzz-testing, a significant amount of time was spend parsing the ROM
database. And most of these tests don't actually use the information from this
database. This patch changes the code so that the ROM database is only created
on first use.

In more realistic scenarios (non fuzz-testing) the ROM database is more likely
to be needed. Though lazily creating it also doesn't hurt.