Added SSE2 version of Sha1Sum::parse40()
After the previous two patches, the next most expensive function during
fuzz-testing was (surprisingly) Sha1Sum::parse40(). This is a relatively simple
function that parses a 40-character hex-string into a 160-bit binary number.
During start-up this function is called many times when parsing the .filecache
file.

The original code had quite a few branches per character. Sha1 values are hard
to distinguish from random numbers. And therefor the branches (that e.g. test
on the char ranges 0-9,a-f,A-F) are also hard to predict. I suspect this is the
main reason why this function was so slow.

It is possible to turn this into branch-less code. Though that's only
profitable when multiple characters can be processed at once, for example by
using SIMD. And that's what this patch does: it adds a SSE2 version of this
function. SSE registers are 128-bit wide, so most operations work on 16
characters at-a-time.

The new SSE2 version is about 4x-5x faster than the plain C++ version. Though
specifically in the context of afl-fuzz the difference is even bigger (factor
10x) because there we use an instrumented openmsx executable where each branch
has some additional overhead.