Move initialization of VDPAccessSlots from run-time to compile-time
C++14 constexpr allows to move relatively complex calculations from run-time to
compile-time (of course given that all inputs for the calculation are already
known at compile-time). This patch uses that feature to calculate the content
of the VDPAccessSlots lookup tables at compile-time. This saves 5% startup cost
when using afl-fuzz.

I hacked our build setup to conditionally pass '-std=c++14' when the gcc
version supports it (to overrule the '-std=c++11' flag which is passed
unconditionally). Though this is unnecessarily complex. It would be better to
simply require c++14 (and thus drop support for gcc-4.8 and gcc-4.9). But ATM
the Android build still uses gcc-4.8. We should upgrade that soonish.