Faster FilePool parsing (mostly by lazy date parsing)
Before this patch over 15% of the total startup time was spend in parsing the
'.filecache' file. And 3/4 of that was in parsing the date/time string to a
time_t value. Another significant part was copying the filenames from the file
to std::string objects. This patch addresses both issues, and now '.filecache'
parsing is over 10x faster!

Parsing the date/time string is now done lazily. Of all the .filecache entries
(several 1000's) only a few will actually be used, and only for those do we
need to verify the stored date/time with the current file modification
timestamp. So only for those we actually need to parse the date/time string.

Though to be able to parse the date/time on demand we need to remember the
corresponding string. Making a copy of all those strings would also be a big
cost. What we do instead is keep the content of the .filecache file in memory,
and then refer to the strings in that memory block. We do the same for the
filenames in that block to also avoid the cost of copying those strings.