Sort 'keys' array during compile-time
About 0.5% of startup time was spend on sorting the 'keys' lookup table. The
content of that table is completely known at compile time, so there's no reason
why it should be sorted during runtime rather than already during compile time.

This patch is relatively large because:
- It's not yet possible to use std::sort() for compile-time calculations (IOW
  std::sort() is not yet constexpr). Though there are proposals to change that
  in the next version of C++. And in fact large parts of this patch can already
  be removed when we switch to C++17.
- It contains fallback code for C++11 compiler (rather than C++14). Once we
  drop gcc-4.8 that fallback code can be removed.

Apart from compile-time sorting this patch also changes the type of the table
from a std::vector to a std::array. This removes the need for dynamic memory
allocation and it also removes one indirection when using the table. So apart
from faster startup it also slightly improves the performance when using the
table.