Move modifier constants from Keyboard to KeyInfo
In the future I want to remove the hardcoded layout of modifier keys,
which means that the keyboard matrix positions of the modifier keys
will no longer be constants. In the Unicode map we still need to
identify parsed modifiers though, so I moved the constants there.

Note that this commit does not yet remove the assumption that
all modifiers are in fixed positions in row 6, it only prepares
the code for later removal of that assumption.