Wrap keyboard matrix positions in a class
Often row and column are passed together to a function, so having
a combined type for them makes sense. And in those cases where they
are not passed together, that is typically the result of assumptions
such as modifiers being on the same row, which are not valid for
non-MSX keyboards.

Instead of testing position and KeyInfo validity in a number of
different ways, we now have explicit isValid() methods.

This also removes the reconstruction of the column value from the
mask in needsLockToggle(), as we now store the column and construct
the mask from that on demand.