Disallow removing non-existing characters from start/end of string_ref
This should be more efficient because an "if" statement becomes
obsolete. Also it is consistent with the std::string_view class
that will eventually replace string_ref.

I checked whether any code depended on the old, more lenient behavior
and did not find any. I added an assert in the only place where the
the directly preceding code did not guarantee that the remove/pop
methods would not remove non-existing characters (the caller did
provide such a guarantee).