Calculate YMF262 sin and exp tables at compile-time
This is more complex than the previous patches because the mathematical
functions sin(), exp() and log() are not constexpr (gcc offers this as an
extension but clang and visual studio do not). Therefor this patch also
implements approximations for those functions.

These are simple textbook approximations based on Taylor-expansion or
Newton-Raphson-iteration. I only checked that these approximation converge for
the domain required in this specific application. You must manually specify
(via trial and error) the number of required iterations in the approximation. I
did verify that the final result is bit-identical before and after this patch.

I consider this an experimental patch. But if it turns out to be successful we
can later do something similar to also remove some code generators like:
   src/sound/generateYM2413OkazakiTable.cc
   src/sound/generateBlipTable.cc