Check modifier keys one at a time in updateKeyMatrix()
This removes the assumption that all modifiers are in the same
keyboard row from updateKeyMatrix().

Note that previously we didn't keep track of CAPS state. Since CAPS
is always a lock key, the modifier state we track for it will not
be used. But not tracking it would require more effort than tracking
all modifiers.