[2a/6] Signedness bug-fix in Y8950 ra-adjust
The intention was to have a negative value in some intermediate calculation,
but because of a type bug we instead got a big positive value. This resulted in
a floating point overflow and later in undefined behavior. But strangely enough
the final result was nevertheless correct.

The next patch will change these calculations to constexpr. And without this
fix this (rightfully) won't compile: constexpr calculations are not allowed to
contain undefined behavior.