Added (disabled) YMF262 table printing code
I used this to manually verify that the content of the tables remained the same
before and after the constexpr refactoring. E.g. to check that the constexpr
math function approximations are accurate enough.

Could maybe be turned into a more automated test in the future. But for now I
commit what I have.