Created unicodemap files for Sega SC-3000 keyboards
There is a Japanese variant and an international variant. The layout
of the keyboard is the same, but the characters produced are
different, particularly for CODE. Also CODE is kana lock on the
Japanese keyboard and ENG/DIER'S (non-locking) on the international
keyboard. "Locking" refers to how Sega BASIC treats the key; there
is no lock LED.

Thanks to Quibus and hap for figuring out the weekday kanjis.