Improve accuracy of cstd::sin by adding a reduction step
cstd::sin() is our compile-time (=constexpr) version of std::sin(). The gcc
version of std::sin() is already constexpr, but the clang and visual studio
versions are not. It's also not (yet?) required by the c++ standard.

Until now we only calculated sine values for the interval [0, pi/2]. But the
next patch requires sine values over a wider domain. The old implementation of
cstd::sin() converged slowly for inputs too far removed from 0. This patch
fixes that by first reducing the input to the interval [0, pi/4] before
calculating the Taylor series.

The next patch also requires cosine, so this patch adds cstd::cos().